Only run Apache if config file exists

+++ apache2.sh
@@ -4,8 +4,12 @@
 mkdir -p /srv/etc
 cp /etc/apache2/conf-available/SOGo.conf /srv/etc/apache-SOGo.conf.orig
 
-# Copy back and enable administrator's configuration
-cp /srv/etc/apache-SOGo.conf /etc/apache2/conf-enabled/SOGo.conf
+# Only run apache if config file exists
+if [ -f /srv/etc/apache-SOGo.conf ] then
 
-# Run apache in foreground
-APACHE_ARGUMENTS="-DNO_DETACH" exec /usr/sbin/apache2ctl start
+    # Copy back and enable administrator's configuration
+    cp /srv/etc/apache-SOGo.conf /etc/apache2/conf-enabled/SOGo.conf
+
+    # Run apache in foreground
+    APACHE_ARGUMENTS="-DNO_DETACH" exec /usr/sbin/apache2ctl start
+fi
